requirements.txt
google-generativeai
gspread
streamlit
oauth2client
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas